Action 微修正

+++ build.sh
@@ -1,5 +1,11 @@
-#!/bin/sh
-rm -rf ./dist
+#!/bin/bash -eu
+
+cd "$(dirname "$0")"
+
+if [ -d ./dist ]; then
+    rm -rf ./dist
+fi
+
 yarn build
 
 cd ./dist
